requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.7
django-extensions==3.1.1
django-heroku==0.3.1
django-phonenumber-field==5.0.0
django-url-or-relative-url-field==0.1.2
Faker==6.6.2
gunicorn==20.0.4
kiwisolver==1.3.1
phonenumbers==8.12.20
psycopg2==2.8.6
py-notifier==0.1.1
pyclean==2.0.0
pydotplus==2.0.2
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
win10toast==0.9
wincertstore==0.2